backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Pillow
celery
redis
gunicorn
pydantic
passlib
pydantic[email]
pytest-flask
Flask-CORS
PyJWT